feat: add platform extensions
