--- packages/components/src/DesktopDragZoneBox/index.native.tsx
@@ -1,4 +1,0 @@
-import { Fragment } from 'react';
-
-export const DesktopDragZoneAbsoluteBar = Fragment;
-export default Fragment;
