--- packages/kit/src/views/Overlay/UpdateItem/index.native.tsx
@@ -1,4 +1,0 @@
-/**
- * UpdateItem component is not needed for native.
- */
-export default () => null;
