+++ packages/kit/src/views/Overlay/showSplashScreen.ios.tsx
@@ -1,0 +1,5 @@
+import { NativeModules } from 'react-native';
+
+export const showSplashScreen = () => {
+  NativeModules.SplashScreenManager.show();
+};
